# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nan Ayeyar API
  description: E-commerce API for rice trading platform
  version: 1.0.0
  contact:
    name: Nan Ayeyar Support
    email: support@nanayeyar.com
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.nanayeyar.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get products
      description: Get paginated list of products with optional locale filtering
      security: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: locale
          in: query
          description: Language preference
          schema:
            type: string
            enum: [en, my]
            default: en
        - name: search
          in: query
          description: Search term
          schema:
            type: string
        - name: outOfStock
          in: query
          description: Filter by stock status
          schema:
            type: boolean
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Get detailed product information
      security: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
        - name: locale
          in: query
          description: Language preference
          schema:
            type: string
            enum: [en, my]
            default: en
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  product:
                    $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Admin product endpoints
  /admin/products:
    post:
      tags:
        - Admin - Products
      summary: Create product
      description: Create a new product (admin only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name_en
                - price
              properties:
                sku:
                  type: string
                  example: RICE-001
                name_en:
                  type: string
                  example: Premium Jasmine Rice
                name_my:
                  type: string
                  example: ဂျ်မင်းဆန်
                description_en:
                  type: string
                  example: High quality jasmine rice
                description_my:
                  type: string
                  example: အရည်အသွေးမြင့်ဂျ်မင်းဆန်
                price:
                  type: number
                  format: decimal
                  example: 2500.00
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                metadata:
                  type: object
                  example:
                    variety: "Jasmine"
                    grade: "Premium"
                    weight: "5kg"
                    harvest_date: "2024-01-15"
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product created successfully
                  product:
                    $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/products/{id}:
    patch:
      tags:
        - Admin - Products
      summary: Update product
      description: Update product information (admin only)
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name_en:
                  type: string
                name_my:
                  type: string
                description_en:
                  type: string
                description_my:
                  type: string
                price:
                  type: number
                  format: decimal
                disabled:
                  type: boolean
                outOfStock:
                  type: boolean
                allowSellWithoutStock:
                  type: boolean
                metadata:
                  type: object
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Product updated successfully
                  product:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        sku:
          type: string
          example: RICE-001
        name:
          type: string
          example: Premium Jasmine Rice
        description:
          type: string
          example: High quality jasmine rice
        images:
          type: array
          items:
            type: string
          example: ["/uploads/rice-1.jpg", "/uploads/rice-2.jpg"]
        price:
          type: number
          format: decimal
          example: 2500.00
        disabled:
          type: boolean
          example: false
        outOfStock:
          type: boolean
          example: false
        allowSellWithoutStock:
          type: boolean
          example: true
        metadata:
          type: object
          example:
            variety: "Jasmine"
            grade: "Premium"
            weight: "5kg"
            harvest_date: "2024-01-15"
        totalStock:
          type: integer
          example: 150
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            stockEntries:
              type: array
              items:
                $ref: '#/components/schemas/StockEntry'
    StockEntry:
      type: object
      properties:
        id:
          type: string
          example: clr1234567890
        productId:
          type: string
          example: clr1234567890
        quantity:
          type: integer
          example: 100
        purchasePrice:
          type: number
          format: decimal
          example: 2000.00
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Validation error"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string